+++ lib/features/rewards/services/reward_service.dart
@@ -1,0 +1,7 @@
+import 'package:payout/features/rewards/models/reward_models.dart';
+
+class RewardService {
+  static double calculateTotalCashback(List<CashbackModel> list) {
+    return list.map((c) => c.amount).fold(0.0, (sum, val) => sum + val);
+  }
+}